Globally replaced "spiritual" with "relig".
(I had used "spiritual" as label for a certain aspect of Balinese
cultural patterns in which there is nothing separate that can be called
"religion"; as in many cultures, it's all religion--farming, insects,
water, whatever.  Maybe this is more true of the elements that are the
focus of the Lansing-Kremer model than in some other cases.  But
"spiritual" may make some researchers/readers/listeners
uncomfortable--it suggests that I'm focusing on something other than
what I'm focusing on--which is what most science-oriented people call
"religion".  It's a better term, and I can always discount it when
talking to e.g. anthropologists who know better.)