trivial. maybe last commit before agentset -> list for communic
Currently, both cropping strategy transmission and relig-type
transmission use agentsets.  (I think the former originally used lists
in Janssen's version, but I eventually replaced it.)  However, I'm
planning on turning this all into list operations.  I might want lists
for relig-type transmission so that I can use sampling with replacement
when the number of speakers is large.  This would probably not really
make a difference in overall behavior, and would simplify the use
of random-poisson to choose how many persons to listen to.
But this means that I would have to change find-best, which means
changing the cropping plan transmission too--or duplicating code
for the two kinds of transmission.